[SmartVersion] Initial phase complete
- Prints version number on startup and non-<kbd>^C</kbd> exit
- Updated useragent string, complies with browser spec
- Updated signature, removed timestamp, added owner

Exports of SockBot (e.g. by downloading a tag from GitHub) bake the version number (if that exact version) and commit information into version.js. It's not possible, at that time, to know the distance from the tag, so (for display purposes) that distance is treated as zero.
If the baked information is unavailable, version.js interrogates command-line Git. This is the preferred modus operandi.
If neither the baked information nor Git's own information are available, version.js uses a hardcoded version number.

Todo:
- Handle <kbd>^C</kbd>
- Add a command to interrogate for version information
- Add an admin command to trigger a remote exit (which may cause the wrapper script to trigger an update)

Also included:
- TRWTF.js literally explodes eslint, added to .eslintignore. Fix your shit, sloosecannon!