feat(Commands): Commands can now be pinned to a particular `this` context
By providing a context when registering the command SockBot will ensure that that command will

always be called with that context, allowing use of the keyword `this` if needed or desired by

programming practices within the plugin providing the command.

Any command that relies on the previously undocumented value of `this` within a command handler will

be broken by this update.